Fixed rule for generating HTTP filter.
Changed the rule for generating httpfilter.h so that it uses the
command line option for bpfa instead of piping bpfa's output to
stdout.  The latter creates an empty (and incorrect) httpfilter.h
file while the former generates no file and assures that the build
will fail again in the same place until the pathname to bpfa is
fixed.